add file exists check

+++ start_download.py
@@ -3,10 +3,10 @@
 import asyncio
 import os
 import time
+import redis
 from threading import Thread
 
 __author__ = 'aniss'
-import redis
 
 
 def redis_push(pool, name, url):
@@ -25,34 +25,26 @@ def start_loop(loop):
 
 
 async def async_download(name, url):
-
     print('Start wget %s...' % name)
-
+    name = name+".apk"
     dir = "download_apks/" + name.split(".")[1][0]
     try:
         if not os.path.isdir(dir):
             os.mkdir(dir)
+        if os.path.isfile(dir+"/"+name):
+            print("Already exists %s" % name)
+            return
         command = 'proxychains wget "%s" -O %s/%s.pkg' % (url, dir, name)
 
         dl = await asyncio.create_subprocess_shell(command)
-        # await download(name, url)
         await dl.wait()
         print('Complete wget %s...' % name)
     except Exception as e:
         print(e)
 
 
-
 if __name__ == '__main__':
-    # pool = redis.ConnectionPool(host='localhost', port=6379, decode_responses=True)
-    # redis_push(pool, name="com.qq.exe",
-    #            url='https://sm.myapp.com/original/Audio/QQMusic_Setup_1598.4685_QMgr-15.9.5.0.exe')
-    # redis_push(pool, name="com.qmusic.exe", url='https://sm.myapp.com/original/im/QQ9.0.3-9.0.3.23743.exe')
-    # r = redis.Redis(connection_pool=pool)
-    # r.lrange("apks", 0, -1)
-    # main()
 
-    # print(r.rpop("apks"))
     if not os.path.isdir("download_apks"):
         os.mkdir("download_apks")
     rcon = get_redis()
